Add initial state, data and generation
First, add data, being the base structures and info for races and stats,
which is static information that can't change during a run.

Second, add state, which for now is just the list of existant
characters. This is the information which defines the current app state.
The character does already include the concept of history, which can
currently contain the initial generation of the primary stats, the
initial generation of the secondary stats, and applying an advance.
Shallya's Mercy is part of the primary stat roll data.

Finally, add the generate function, which builds a basic character, so
far only taking a race and rolling initial stats.