Add initial career data
This data still needs to be massaged a bit. Skill, talent, trapping,
entry and exit data is all there, but it's not all in a usable format
yet. The name, quote, source book, description and advances should all
be correct, however. And all the relevant extra rules are in a quote at
the bottom. A subset of that data will need to be displayed in the
app eventually, but it needs more parsing.