Put real info in the history bar
Note that none of this is interactive yet. It just displays info.

- Removed the CREATION section of the bar, moving everything that would
go in there to the top of the HISTORY section.
- Remove skillChoices from Character, instead encoding that as a set of
automaticly generated HistorySkillAdvance events in the history.
- Add getAvailableSkills to character.ts, to get the list of all skills
that the character could currently purchase.
- Put Shallya's Mercy in the history bar, correctly listing on those
stats which rolled below 11, and showing their before and after values.
- Put all HistorySkillAdvance and HistoryStatAdvance events in the
history bar.
- Put all skills available for purchase in the ADVANCES section of the
history bar.