Fix but in removing history events
It would previously always remove the last event when you tried to
remove any event. This turned out to be due to a bug in closing over a
number in a loop, so every closure ended up with the final state of the
number, at the end of the loop.